integrate mount.rs into conn.rs
